# app/lib/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample video tutorial",
    "version": "1.0.0",
    "title": "Tutorial",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "tutorial@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/",
...
  "schemes": ["http", "https"],
  "paths": {
...
    "/auth/forgotPassword": {
      "post": {
        "tags": ["Auth"],
        "summary": "Forgot Password",
        "description": "Quên mật khẩu",
        "operationId": "forgotPassword",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/forgotPassword"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/auth/resetPassword/{resetToken}": {
      "post": {
        "tags": ["Auth"],
        "summary": "Resset Password",
        "description": "Đổi mật khẩu",
        "operationId": "resetPassword",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "resetToken",
            "description": "resetToken of body",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created user Record",
            "required": true,
            "schema": {
              "$ref": "#/definitions/resetPassword"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "forgotPassword": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "resetPassword": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        }
      }
    },
...
  },
...
}